requirements.txt
aspose_cells==23.7.0
firebase_admin==6.2.0
JPype1==1.4.1
JPype1==1.4.1
pandas==2.0.3
plotly==5.15.0
PySimpleGUI==4.60.5
streamlit==1.25.0
streamlit_autorefresh==1.0.1
